+++ test/index.js
@@ -9,9 +9,9 @@ var assert = require('chai').assert;
 var bunyan = require('bunyan');
 var crypto = require('crypto');
 var format = require('util').format;
+var restify = require('restify');
 var uuid   = require('uuid');
 
-var restify = require('restify');
 var clients = require('../lib');
 var auditor = require('../lib/helpers/auditor');
 var pkgJson = require('../package');
@@ -22,7 +22,6 @@ var PORT = process.env.UNIT_TEST_PORT || 0;
 var JSON_CLIENT;
 var STR_CLIENT;
 var RAW_CLIENT;
-var TIMEOUT_CLIENT;
 var SAFE_STRINGIFY_CLIENT;
 var SERVER;
 
@@ -112,28 +111,19 @@ function sendSignature(req, res, next) {
     }
 }
 
-
 function sendWhitespace(req, res, next) {
-    var body = ' ';
+    // override contentType as otherwise the string is json-ified to
+    // include quotes. Don't want that for this test.
+    var count = parseInt(req.params.count);
+    var whitespaces = '';
 
-    if (req.params.flavor === 'spaces') {
-        body = '   ';
-    } else if (req.params.flavor === 'tabs') {
-        body = ' \t\t  ';
+    while (count--) {
+        whitespaces += ' ';
     }
 
-    // override contentType as otherwise the string is json-ified to
-    // include quotes. Don't want that for this tesassert.
     res.header('content-type', 'text/plain');
-    res.send(body);
+    res.send(whitespaces);
     next();
-}
-
-function requestThatTimesOut(req, res, next) {
-    setTimeout(function () {
-        res.send('OK');
-        next();
-    }, 170);
 }
 
 function sendJsonZero(req, res, next) {
@@ -148,10 +138,29 @@ function sendJsonFalse(req, res, next) {
     next();
 }
 
+function sendJsonInvalid(req, res, next) {
+    res.header('content-type', 'application/json');
+    res.end('{"foo":"bar}');
+    return next();
+}
+
+function sendJsonInvalid500(req, res, next) {
+    res.header('content-type', 'application/json');
+    res.status(500);
+    res.end('{"foo":"bar}');
+    return next();
+}
+
 function sendJsonNull(req, res, next) {
     res.header('content-type', 'json');
     res.send(200, null);
     next();
+}
+
+function sendJsonString(req, res, next) {
+    res.header('content-type', 'application/json');
+    res.send({ hello: req.params });
+    return next();
 }
 
 function getLog(name, stream, level) {
@@ -184,7 +193,8 @@ describe('restify-client tests', function () {
         try {
             SERVER = restify.createServer({
                 dtrace: dtrace,
-                log: getLog('server')
+                log: getLog('server'),
+                handleUncaughtExceptions: false
             });
 
             SERVER.use(restify.acceptParser(['json', 'text/plain']));
@@ -195,7 +205,7 @@ describe('restify-client tests', function () {
             SERVER.use(restify.bodyParser());
 
             SERVER.get('/signed', sendSignature);
-            SERVER.get('/whitespace/:flavor', sendWhitespace);
+            SERVER.get('/whitespace/:count', sendWhitespace);
 
             SERVER.get('/json/boom', function (req, res, next) {
                 res.set('content-type', 'text/html');
@@ -214,6 +224,9 @@ describe('restify-client tests', function () {
             SERVER.get('/json/zero', sendJsonZero);
             SERVER.get('/json/false', sendJsonFalse);
             SERVER.get('/json/null', sendJsonNull);
+            SERVER.get('/json/invalid', sendJsonInvalid);
+            SERVER.get('/json/invalid500', sendJsonInvalid500);
+            SERVER.post('/json/string', sendJsonString);
 
             SERVER.get('/json/:name', sendJson);
             SERVER.head('/json/:name', sendJson);
@@ -222,8 +235,6 @@ describe('restify-client tests', function () {
             SERVER.patch('/json/:name', sendJson);
             SERVER.del('/json/:name', sendJson);
             SERVER.opts('/json/:name', sendJson);
-
-            SERVER.get('/str/request_timeout', requestThatTimesOut);
 
             SERVER.del('/str/:name', sendText);
             SERVER.get('/str/:name', sendText);
@@ -264,11 +275,6 @@ describe('restify-client tests', function () {
                         accept: 'text/plain'
                     }
                 });
-                TIMEOUT_CLIENT = clients.createStringClient({
-                    url: 'http://127.0.0.1:' + PORT,
-                    requestTimeout: 150,
-                    retry: false
-                });
                 SAFE_STRINGIFY_CLIENT = clients.createJsonClient({
                     url: 'http://127.0.0.1:' + PORT,
                     dtrace: dtrace,
@@ -291,7 +297,6 @@ describe('restify-client tests', function () {
             JSON_CLIENT.close();
             STR_CLIENT.close();
             RAW_CLIENT.close();
-            TIMEOUT_CLIENT.close();
             SAFE_STRINGIFY_CLIENT.close();
             SERVER.close(callback);
         } catch (e) {
@@ -300,6 +305,56 @@ describe('restify-client tests', function () {
         }
     });
 
+    describe('req.getTimings()', function () {
+        var TIMINGS_CLIENT_IP;
+        var TIMINGS_CLIENT_HOST;
+
+        beforeEach(function () {
+            TIMINGS_CLIENT_IP = clients.createJsonClient({
+                url: 'http://127.0.0.1:' + PORT,
+                retry: false
+            });
+            TIMINGS_CLIENT_HOST = clients.createJsonClient({
+                url: 'http://localhost:' + PORT,
+                retry: false
+            });
+        });
+
+        afterEach(function () {
+            TIMINGS_CLIENT_IP.close();
+            TIMINGS_CLIENT_HOST.close();
+        });
+
+        it('returns timings with IP', function (done) {
+            TIMINGS_CLIENT_IP.get('/json/mcavage', function (err, req, res) {
+                assert.ifError(err);
+                var timings = req.getTimings();
+                assert.isObject(timings);
+                assert.isNull(timings.dnsLookup);
+                assert.isNull(timings.tlsHandshake);
+                assert.isNumber(timings.tcpConnection);
+                assert.isNumber(timings.firstByte);
+                assert.isNumber(timings.contentTransfer);
+                assert.isNumber(timings.total);
+                done();
+            });
+        });
+
+        it('returns timings with Host', function (done) {
+            TIMINGS_CLIENT_HOST.get('/json/mcavage', function (err, req, res) {
+                assert.ifError(err);
+                var timings = req.getTimings();
+                assert.isObject(timings);
+                assert.isNumber(timings.dnsLookup);
+                assert.isNull(timings.tlsHandshake);
+                assert.isNumber(timings.tcpConnection);
+                assert.isNumber(timings.firstByte);
+                assert.isNumber(timings.contentTransfer);
+                assert.isNumber(timings.total);
+                done();
+            });
+        });
+    });
 
     it('GET json', function (done) {
         JSON_CLIENT.get('/json/mcavage', function (err, req, res, obj) {
@@ -307,31 +362,6 @@ describe('restify-client tests', function () {
             assert.ok(req);
             assert.ok(res);
             assert.deepEqual(obj, {hello: 'mcavage'});
-            done();
-        });
-    });
-
-    it('GH-778 GET jsonp', function (done) {
-        // Using variables here to keep lines under 80 chars
-        var jsonpUrl = '/json/jsonp?callback=testCallback';
-        var expectedResult = 'typeof testCallback === \'function\' && ' +
-                             'testCallback({"hello":"jsonp"});';
-
-        JSON_CLIENT.get(jsonpUrl, function (err, req, res) {
-            assert.ifError(err);
-            assert.ok(req);
-            assert.ok(res);
-            assert.equal(res.body, expectedResult);
-            done();
-        });
-    });
-
-    it('GH-388 GET json, but really HTML', function (done) {
-        JSON_CLIENT.get('/json/boom', function (err, req, res, obj) {
-            assert.ifError(err);
-            assert.ok(req);
-            assert.ok(res);
-            assert.deepEqual(obj, {});
             done();
         });
     });
@@ -347,7 +377,7 @@ describe('restify-client tests', function () {
             });
         } catch (err) {
             assert.ok(err);
-            assert.equal(err.constructor, TypeError);
+            assert.equal(err instanceof TypeError, true);
             var errMsgRe = /^Request path contains unescaped characters\.?$/;
             assert.ok(errMsgRe.test(err.message),
                 format('error message matches %s: %j', errMsgRe, err.message));
@@ -459,8 +489,8 @@ describe('restify-client tests', function () {
             assert.ifError(err);
             assert.ok(req);
             assert.ok(res);
-            assert.deepEqual(obj, {});
             assert.strictEqual(res.body, '0');
+            assert.strictEqual(obj, '0');
             done();
         });
     });
@@ -471,8 +501,8 @@ describe('restify-client tests', function () {
             assert.ifError(err);
             assert.ok(req);
             assert.ok(res);
-            assert.deepEqual(obj, {});
             assert.strictEqual(res.body, 'false');
+            assert.strictEqual(obj, 'false');
             done();
         });
     });
@@ -483,8 +513,8 @@ describe('restify-client tests', function () {
             assert.ifError(err);
             assert.ok(req);
             assert.ok(res);
-            assert.deepEqual(obj, {});
             assert.strictEqual(res.body, 'null');
+            assert.strictEqual(obj, 'null');
             done();
         });
     });
@@ -734,32 +764,6 @@ describe('restify-client tests', function () {
         });
     });
 
-    it('GH-20 connectTimeout', function (done) {
-        var client = clients.createClient({
-            url: 'http://169.254.1.10',
-            type: 'http',
-            accept: 'text/plain',
-            connectTimeout: 100,
-            retry: false,
-            agent: false
-        });
-
-        client.get('/foo', function (err, req) {
-            assert.ok(err);
-            assert.equal(err.name, 'ConnectTimeoutError');
-            done();
-        });
-    });
-
-    it('requestTimeout', function (done) {
-        TIMEOUT_CLIENT.get('/str/request_timeout',
-            function (err, req, res, obj) {
-            assert.ok(err);
-            assert.equal(err.name, 'RequestTimeoutError');
-            done();
-        });
-    });
-
     it('GH-169 PUT json Content-MD5', function (done) {
         var msg = {
             _id: '4ff71172bc148900000010a3',
@@ -848,28 +852,6 @@ describe('restify-client tests', function () {
             assert.ifError(err);
             assert.ok(req);
             assert.ok(res);
-            done();
-        });
-    });
-
-
-    it('GH-203 GET json, body is whitespace', function (done) {
-        JSON_CLIENT.get('/whitespace/spaces', function (err, req, res, obj) {
-            assert.ifError(err);
-            assert.ok(req);
-            assert.ok(res);
-            assert.deepEqual(obj, {});
-            done();
-        });
-    });
-
-
-    it('GH-203 GET json, body is tabs', function (done) {
-        JSON_CLIENT.get('/whitespace/tabs', function (err, req, res, obj) {
-            assert.ifError(err);
-            assert.ok(req);
-            assert.ok(res);
-            assert.deepEqual(obj, {});
             done();
         });
     });
@@ -1192,6 +1174,82 @@ describe('restify-client tests', function () {
             assert.equal(err.name, 'TooManyRedirectsError');
             client.close();
             done();
+        });
+    });
+
+    it('GH-147: JSONClient should errback on receiving invalid JSON',
+    function (done) {
+        JSON_CLIENT.get('/json/invalid', function (err, req, res, data) {
+            assert.ok(err);
+            assert.deepEqual(err.name, 'RestError');
+            assert.deepEqual(err.message, 'Invalid JSON in response');
+            assert.deepEqual(err.cause().name, 'SyntaxError');
+            assert.include(err.cause().message, 'Unexpected end of');
+            assert.equal(200, res.statusCode);
+            assert.equal(data, '{"foo":"bar}');
+            assert.equal(res.body, '{"foo":"bar}');
+            return done();
+        });
+    });
+
+    it('GH-147: JSONClient should prefer http error over parse error',
+    function (done) {
+        JSON_CLIENT.get('/json/invalid500', function (err, req, res, data) {
+            assert.ok(err);
+            assert.deepEqual(err.name, 'InternalServerError');
+            // should include original unparsed JSON payload
+            assert.include(err.message, '{"foo":"bar}');
+            assert.isUndefined(err.cause());
+            assert.deepEqual(res.statusCode, 500);
+            assert.deepEqual(data, '{"foo":"bar}');
+            assert.deepEqual(res.body, '{"foo":"bar}');
+            return done();
+        });
+    });
+
+    it('GH-147: JSONClient should not error on receiving txt/plain',
+    function (done) {
+        JSON_CLIENT.get('/str/foobar', function (err, req, res, data) {
+            assert.ifError(err);
+            assert.strictEqual(res.statusCode, 200);
+            assert.strictEqual(data, 'hello foobar');
+            return done();
+        });
+    });
+
+    it('GH-147: JSONClient should accept string input', function (done) {
+        JSON_CLIENT.post('/json/string', 'foobar',
+        function (err, req, res, data) {
+            assert.ifError(err);
+            assert.strictEqual(res.statusCode, 200);
+            // raw res.body is unparsed JSON
+            assert.deepEqual(res.body, '{"hello":"foobar"}');
+            assert.deepEqual(data, { hello : 'foobar' });
+            return done();
+        });
+    });
+
+    it('GH-147: whitespace should return parse error', function (done) {
+        JSON_CLIENT.get('/whitespace/1', function (err, req, res, data) {
+            assert.ok(err);
+            assert.deepEqual(err.name, 'RestError');
+            assert.deepEqual(err.message, 'Invalid JSON in response');
+            assert.deepEqual(err.cause().name, 'SyntaxError');
+            assert.include(err.cause().message, 'Unexpected end of');
+            assert.strictEqual(res.statusCode, 200);
+            // raw res.body is unparsed JSON
+            assert.deepEqual(res.body, ' ');
+            assert.deepEqual(data, ' ');
+            return done();
+        });
+    });
+
+    it('GH-147: empty payload should return empty pojo', function (done) {
+        JSON_CLIENT.get('/whitespace/0', function (err, req, res, data) {
+            assert.ifError(err);
+            assert.deepEqual(res.body, '');
+            assert.deepEqual(data, {});
+            return done();
         });
     });